+++ src/gui.cpp
@@ -1332,7 +1332,7 @@ bool gui_textfield(GuiContext *ctx, const char *label, char *buf, int buf_size)
 
         if (active) {
             assert(buf && buf_size > 0);
-            int char_count = strlen(buf);
+            int char_count = static_cast<int>(strlen(buf));
             for (int i = 0; i < ctx->written_char_count; ++i) {
                 if (char_count >= buf_size)
                     break;
